Don't call clean with `-x`
This would remove all files from the checkout (including those in the
.gitignore list) but this is a pretty big change in behaviour that can
use alot of things to break. If users are expecting a file to persist
between runs, then this would break that.

If someone wants to do a `git clean -x`, they'll be able to use the hook
system that will be in 1.0, or use the CLEAN_CHECKOUT gear.