Store publicly-accessible URL for Google Cloud Storage
When an artifact is stored on GS, it can be downloaded from the
Buildkite UI if it's publicly accessible. Currently, we store URLs of
the form:

    https://www.googleapis.com/storage/v1/b/<bucket>/o/<object>?alt=media

This is problematic because `<object>` has slashes escaped, so
downloading `foo/bar/blah.txt` results in a file `foo-bar-blah.txt`.

Fix this by using the form:

    https://storage.googleapis.com/<bucket>/<object>

This doesn't require escaping slashes, so downloading an object
`foo/bar/blah.txt` results in a file `blah.txt`, as expected.

Note: There is yet another form of URL:

    https://storage.cloud.google.com/<bucket>/<object>

This URL always requires cookie-based authentication, and will
redirect through an OAuth flow if necessary. I considered using this
form, but decided against it because it *always* requires a valid
cookie, even for publicly-accessible objects.