bootstrap: Add support for BUILDKITE_REFSPEC
This is useful to integrating Buildkite with Gerrit. Basically, in the
Gerrit workflow commits for patchsets are not done on branches.
Instead, they each have a ref of the form:

    refs/changes/XX/XX/XX

In order to build the commit, you must first pull the ref. So add
support for a new environment variable BUILDKITE_REFSPEC that can be
set to indicate the refspec that Buildkite should pull.

This is a generalization of the support for building pull requests
from GitHub. In that case Buildkite figures out the ref automatically
from GITHUB_PULL_REQUEST environment variable.