bootstrap: Make "git clean" behavior configurable
The default `git clean` flags are `-fdq`. It's useful to be able to
configure this. For example, `-x` can be used to remove gitignore'd
files. Or `-e` can be used to exclude specific directories.

Introduce an environment variable `BUILDKITE_GIT_CLEAN_FLAGS` to
control this behavior. This can then be set in an `environment` hook.

Closes #173