Don't presume hashed known hosts
We were checking for a hashed known host only, but we were scanning an
unhashed host so the check would never become true.

Instead check if we have the unhashed host, or the hashed host (which
has probably been manually added), otherwise scan the unhashed host
which is compatible with both potential configurations.