Disable CGO
We really want to ship an entirely static binary to make it as easy as
possible for folks to run the buildkite-agent without worrying about
glibc etc. So let's disable CGO entirely.

This means the system's networking configuration may not be entirely
honoured. On *nix systems Go will read /etc/resolv.conf and use those
dns servers, but otherwise skip any other name resolution mechanisms
like mDNS. This includes losing Bonjour/mDNS on macOS.

At the moment this also means we lose process titles, but they've only
been available on amd64 linux until now and we have a pure Go solution
to that in the works.

For Windows, this shouldn't cause any change.