Upload github release and homebrew in beta/stable pipelines
Previously we were doing the github release and homebrew release only in
the beta pipeline, which meant that when we released 3.0.0 it shipped
even before the stable release pipeline was unlocked, which was very
unexpected (for me).

This looks at the pipeline and skips the github and homebrew step
depending on whether the version is a pre-release or not.