edit Buildkite Agent license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the LICENSE file so that it contains only the full
text of the MIT license. It also moves the info pertaining to 3rd-party
dependencies to a new DEPENDENCIES.txt file and makes note of this new
file in the README's "Copyright" section.

Collectively, these changes allow Licensee to successfully identify the
license type of Buildkite Agent as MIT.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>