Use .gitmodules for determining what hosts to keyscan
Currently we use a git submodule foreach to try and get the url of all
our submodules so that we can figure out if any need ssh keyscanning.
This is unfortunately completely flawed as none of them exist until
after we call git submodule update.

This uses the git config command to parse the .gitmodule file instead.