Don't force set the executable bit on scripts to be executed
At present, buildkite sets the executable bit on any script it attempts to
execute as part of a job.

This is actually quite hazardous as there may be files that are valid
interpreter scripts, but are dangerous to run -- by having this behaviour, it
actually subverts any good that no-command-eval can achieve by leaving another
potentially gaping hole which we can't turn off!

It's also relatively unnecessary - git, by default, tries to store the state of
the execute bit into the repo.  We should not subvert this if the user doesn't
want us to.

This patch disables this behaviour when CommandEval is also disabled.