Added feature to enable encryption at rest for artifacts in S3.
This uses a new ENV variable BUILDKITE_S3_SSE_ENABLED in line with the
existing enviroment variables.

Currently validate of this value is optimised for feedback, being limited
to covering the case of encryption at rest. If this accepted the full
range of options supported by S3 it would be difficult to validate
as it requires a how load of KMS infrastructure and permissions.

Also extracted the existing flag and wrote some tests for it and the new
option.