Propagate signals in intermediate bash shells
If you execute multiple commands, for instance [a, b, c] then we join
them together in a single bash shell. This has the side effect of
creating and intermediate bash shell where you might not expect it. This
adds a trap that propagates signals to the process group when it gets
them.

This means that cancel signals propagate correctly.