Download absolute-path artifacts to correct path
The handling of artifacts uploaded with absolute paths is currently
undefined and resulted in it being handled as a relative path instead.

We add explicit handling of absolute paths. The download configuration
of such artifacts still requires a download path to be specified as the
nature of paths of artifacts that will end up being downloaded can only
be determined at runtime.

We also add a new test to prevent regression of the newly implemented
behaviour as well explicit documentation in the CLI help.

Fixes #1035.