Alter debian package build step to work on agents without ruby
This step requires an environment with ruby (to use the fpm gem).
Historically we've used the "queue=deploy" targeting to ensure it runs
on an agent that happens to have ruby.

However, we'd like to be able to run this step on a standard elastic
stack environment, which only has the bk agent binary, awscli, and
docker.

This changes the step to launch a docker container with ruby for
the parts that need it.

I considered using the docker-compose plugin for the whole step like
other parts of the pipeline, however there's a reasonable amount of
existing shell code used here and I wasn't sure how it would interact
with the plugin. In the end, I optimised for the least disruption to
what was already there.