run "publish to s3" steps on the new deploy-production agents
The "deploy" queue currently runs on the bastion, and we're trying to
move things off that so it can be shut down.

publishing to s3 requires the buildkite-agent (to fetch metadata) and
awscli (to push to s3), both of which are available on the elastic stack
agents that process the "deploy-production" queue.

There's a good chance the agents won't have IAM permissions for pushing
to the bucket, but we can address that if it happens.