publish-to-s3: use a docker container to execute ruby
The new agents this step runs on don't have ruby available, so a docker
container is our easiest way forward.

The output of this script is significant, so I had to remove some stdout
output from the ruby-env script. I think that's OK - nothing else should
be depending on it.