stage docker images in ECR rather than docker hub
Historically, we've used a private docker hub area to hold images that
may eventually be released as public images on [1].

These days we're using ECR a lot more though, and managing
authentication is much easier. It's one less secret for us to handle.

The agents that currently run this step don't auto-login to ECR, so we
need to do so manually.

[1] https://hub.docker.com/r/buildkite/agent