publish images to docker hub using the deploy queue
The deploy-legacy queue is deprecated, and we're slowly moving steps off
it.

There's two parts to publishing:

1. pulling the staged images from ECR
2. pushing the images to docker hub

For (1), the agents processing the new deploy queue are an elastic
stack, and the instance roles in use have permission to pull from ECR.
They also can use plugins, so we can use the ECR plugin to handle logins
for us.

For (2), we need to manually pull the credentials from SSM and call
"docker login". The agent that used to run this step had a hook that
handled docker hub authentication, but that's not our preferred approach
any more.