move homebrew releases from the deploy-legacy queue, to the deploy queue
The deploy-legacy queue is deprecated and we're slowly moving steps to
the shiny new "deploy" queue.

The new queue is processed by agents in an elastic stack, so we need to
use docker to gain access to the tools we need (like ruby).

Also set the Github API token explicitly. On previous agents this was
set in an agent environment hook. However the agents we want to move
this step to don't set secrets in hooks, so we have to read it from SSM
ourselves