allow beta/stable releases in dry run mode to continue if changelog is missing
This build step helps ensure we keep the changelog up to date. However
in dry run mode we often want to confirm later  steps in the build are
working and we know that areal release isn't happening.

In that case, we can make the changelog check a warning, rather than a
fatal error.