move github releases from the deploy-legacy queue, to the deploy queue
The deploy-legacy queue is deprecated and we're slowly moving steps to
the shiny new "deploy" queue.

The new queue is processed by agents in an elastic stack, so we need to
use docker to gain access to the tools we need (like github-release)

The old agents also set the GITHUB_RELEASE_ACCESS_TOKEN env var in an
agent hook. The new agents don't do that, so we fetch it ourselves from
SSM.