Accept `--experiment` flags in all buildkite-agent subcommands
While the upload path normalisation logic added in #1211 worked in unit tests, it turns out the subcommand wasn't actually accepting the experiment flag, only `buildkite-agent start` was. This updates all other subcommands to read the experiment values, even if they don't themselves implement an experiment. This permits experiments such as that added in #1211 to exist within in-job commands such as `artifact upload`.