update aws-sdk-go to the latest version
We were using a version publishied nearly 2 years ago. We really only
use the s3 support and I can't see any critical updates or bugfixes in
that package, but no harm in trying the update anyway. The CHANGELOG
lists a number of fixes for rare bugs in session handling and retry
logic.