submit basic OS info when registering from a BSD system
On macos, windows and linux the agent registration request includes some
basic info on the operating system. Version, distribution, etc. This
data is used only for metrics on what environments the agent is used in.

We publish official binaries for some BSD systems (FreeBSD, OpenBSD,
NetBSD and dragonfly), however we haven't been submitting any OS
versions so we have a poor understanding of how often these operating
systems are used.

This change includes the output of `uname -a` in agent registration
requests on those operating systems.