read supported environment variable keys in a ufave/cli/v2 compatible way
This method dynamically fetches a list of all the environment variables
that can be used to configure the agent, and removes them from the
environment so they're not passed down to child processes (like the
bootstrap used to run a job).

In ufave/cli/v1 each flag supported a single environment variable key,
stored in the EnvVar field. In ufave/cli/v2 each flag can have multiple
environment variable keys and they're stored in a slice of Strings
called EnvVars (note the `S`).

This now reads the appropriate field, and does some sanity checking to
ensure the method fails loudly if ufave/cli/v3 decides to change/rename
things.