Enable BUILDKITE_AGENT_DEBUG_HTTP in jobs if it's enable in the agent process
BUILDKITE_AGENT_DEBUG is already passed down to the job environment, and
this expands that behaviour to also include BUILDKITE_AGENT_DEBUG_HTTP.

This is desirable because the build-agent binary is often used with jobs
(eg. for artifact uploading/downloading) and from time to time it can be
helpful to see precisely what those HTTP requests look like.