log the S3 IP used during artifact uploads
We get occasional reports of slow and/or failed S3 uploads. One
possibility is that there's networking issues between the agent and
us-east-1 (where the default buildkite-managed artifacts bucket lives).

The IP address for s3.amazonaws.com rotates every 5 seconds, so from
upload to upload it's impossible for us to guess the server IP that was
used after the fact. This will output a log line for each upload (when
BUILDKITE_AGENT_DEBUG_HTTP=true) with the exact server IP address that
was used.

If we see upload issues in the future that sound like a networking
issue, at least we'll an address to debug with.