fail fast when uploading an artifact > 5Gb to the default s3 bucket
Buildkite provides a default artifact storage option that happens to be
stored on S3. Each artifact is uploaded as a signed POST, which has a
limit of 5Gb.

At the moment, artifacts that exceed this limit are uploaded 10 times
(thanks to retries). On each attempt the full file contents is uploaded,
only to be rejected by S3 at the end. What a waste of bandwidth and
time.

This change ensures artifacts that exceed 5Gb display an error before
starting the file upload. This isn't a perfect solution, as the retry
logic kicks in and there's 10 retries. They complete quickly though, and
no bandwidth is wasted.

The error message is long, but I thought it might be helpful to point to
the other artifact storage options we support (like customer owned S3
buckets) where artifacts > 5Gb may work fine.

Closes #746