wip: pipe commit info into buildkite agent meta-data set command
If the commit info is too large then this will fail with an "argument
list too long" error. We can pipe the commit it to fix this, and it
makes it more reliable. But I can't figure out how to punch a hole the
right size through all of these layers of abstraction.

This almost works, but it blocks forever because process.Run() blocks
until the process exits, but it can't exit until we've written and
closed stdin, which we can't do until we start the process in
process.Run().