experiments: local var name refactor
This is mostly to make the implementation of Enabled() clearer.

Idiomatic Go uses `ok` to indicate whether a key _existed_ in a map,
e.g. `value, ok := experiments["foo"]`.
Whereas we were using it to refer to the boolean _value_ of the map in
the context of a range, i.e. `for key, ok := range experiments`.

And `var enabled []string` as the accumulator list of keys which are
enabled is problematic because it isn't plural, and the same `enabled`
name could also refer to a boolean value in the `map[string]bool`. So
this patch renames the list to `keys`, and also uses `key` instead of
`experiment` everywhere to refer to experiment keys.

With the name `enabled` no longer being used for a list of enabled keys,
this patch uses `enabled` instead of `ok`, resulting in a clearer
`if enabled { ... }`.