Downgrade github.com/urfave/cli to v1
In #1233 we upgraded github.com/urfave/cli from v1 to v2. There was no
functional need for the upgrade, it was just about bumping our
dependencies and continuing to slowly adopt go modules.

However, I overlooked that v2 changed some rules for ordering of flags
and arguments. From the [upgrade
guide](https://github.com/urfave/cli/blob/master/docs/migrate-v1-to-v2.md):

> In v2 flags must come before args. This is more POSIX-compliant. You may need to update scripts, user documentation, etc.
>
> This will work:
>
>     cli hello --shout rick
>
> This will not:
>
>     cli hello rick --shout

We're not keen on breaking our users scripts, so let's rollback to v1.
There's an [open issue
upstream](https://github.com/urfave/cli/issues/1113) about the new
ordering rules. We'll keep an eye on that, and maybe re-try the upgrade
if it becomes possible to do so in a non-breaking way

This is mostly a revert of PR #1233 and #1250, but with a few extra
changes to things that happened after that PR (like adding the `artifact
search` subcommand).