update github.com/urfave/cli to the latest v1 release
In PR #1233 we upgrade this library to v2, but then reverted it in
PR #1286 after discovering the upgrade lead to a breaking change in our
CLI.

It seems we're stuck on v1 for now, but we may as well update to the
latest v1 release. This takes us to a version with formal go mod
support.