Add a Homebrew-based installation method for macOS hosts
Given that Homebrew no longer supports formula flags with values in them (see Homebrew/brew#7638, buildkite/homebrew-buildkite#18, buildkite/homebrew-buildkite#15), this updates the Unix `install.sh` to allow the use of the Homebrew install method, while also providing one-line token configuration.

To do so, I've divided up the install process into a few more functions, and shared as much of the logic as possible between them.

The original `curl`/`wget`-based installation process is kept for serving other platforms, and is also used if `BETA` is set to `true`, if Homebrew is not found, or if a `DESTINATION` was explicitly specified.

This also currently means that if Homebrew is present and used, our releases endpoint is currently not fetched, so we don't get any statistics. Perhaps we would want this anyway? I guess that's a decision we'd have to make.

I have tested this on macOS, with Homebrew removed from the path, with `DESTINATION` overridden, and with `BETA` set.