Expand the CLI test step
This adds more virtual Buildkite steps to the "bk cli" test. Initially intended for testing the bk CLI against the Agent, this is also a handy way for us to check we haven't broken any Agent CLI arguments.

So, this adds a test which uses the meta-data subcommand with a default set, and a suite of tests which exercise the annotate subcommand.