Use a win32 Job Object to manage job termination on windows
Imagine a windows process tree like this:

    1000 buildkite-agent.exe start
    └─ 1001 buildkite-agent.exe bootstrap
      └─ 1002 cmd.exe
        └─ 1003 python.exe
          └─ 1004 python.exe

It's reasonably common in windows for a process to end and orphan its
children, leaving a tree like this:

    1000 buildkite-agent.exe start
    └─ 1001 buildkite-agent.exe bootstrap
      └─ 1002 cmd.exe
    └─ 1004 python.exe

Then the bootstrap might end, leaving the orphan process hanging around
after the job has completed. Oops.

On UNIX systems we avoid this by creating a progress group, so any
process that starts during job execution is tidied up when the bootstrap
process finishes.

Here I've introduced win32 Job Object to achieve a similar effect on
Windows. This has two noteworthy impacts:

1. The bootstrap process created to run a Buildkite Job is added to a
   new win32 Job Object. Any new process started by descendant
   processes will be added to the same Job Object, and they'll be
   automatically terminated when the bootstrap process finishes
2. A request to hard terminate a running Buildkite Job now does so by
   terminating the win33 Job (which in turn terminates all processes
   within it). This should be more reliable than TASKKILL.EXE, which was
   unable to find and destroy orphans and would error completely if the
   bootstrap process had completed

HT to this gist that pointed me in the right direction [1]. The MSDN
docs on Job Objects [2] were reasonably useful too.

Fixes #1292

[1] https://gist.github.com/hallazzang/76f3970bfc949831808bbebc8ca15209#gistcomment-2948162
[2] https://docs.microsoft.com/en-us/windows/win32/procthread/job-objects?redirectedfrom=MSDN