add sampling logger for job-related processes on windows
We've had reports from multiple customers that the windows agent can
sometimes get "stuck" (eg #1292)

- from the perspective of our servers the agent is running the stuck job for days
- the main agent process considers the job to still be running, so it
  doesn't mark the job as finished with our API
- the main agent process sees no further output from the bootstrap
  process, so no new log chunks are uploaded
- when the user tries to cancel the job on buildkite.com, the main agent
  process attempts to terminate the bootstrap process and fails

This PR doesn't attempt to fix the issue - we're still looking for a
root cause that we can reliably reproduce.

This adds a sampling logger that prints the full process tree of the
bootstrap and job every 5 seconds.

I'm told that the process hierarchy in windows can't be trusted too
much. Each process maintains a pointer to its parent, but it's common
for the parent process to end and leave orphan processes that can't be
found by walking the parent/child tree from the bootstrap.

I've attempted to mitigate that by adding the bootstrap process to a Job
Object. All child processes of the bootstrap are automatically added to
the same Job, and we can use the QueryInformationJobObject windows API
call to retrieve all processes in the Job.

I'm interested to see how the process tree evolves over the life of one
of these stuck jobs. Hopefully we can get the clue we need to reproduce
the issue and come up with a reliable fix.

Note: this is not production ready code! I haven't put much thought into
the abstractions I've modified. To enabled the logging, start the agent
like this:

    buidlkite-agent.exe start -token=foo -debug -experiments=debug-processes

Some useful reading on Job Objects:

1. https://docs.microsoft.com/en-us/windows/win32/procthread/job-objects?redirectedfrom=MSDN
2. https://docs.microsoft.com/en-us/windows/win32/api/jobapi2/
3. https://nikhilism.com/post/2017/windows-job-objects-process-tree-management/
4. https://gist.github.com/hallazzang/76f3970bfc949831808bbebc8ca15209