use golang.org/x/sys/windows instead of syscall, where we can
the syscall package in the stdlib has a frozen interface, however
golang.org/x/sys continues to evolve. Some of the windows APIs that
we're manually accessing via syscall now have go-ish wrappers in
sys/windows that are a bit nicer to work with.

There should be no functional difference from this change.