avoid nil entries when converting PIDs to a Process
We fetch the PIDs in a Job Object, then sequentially lookup the details
of the process associated with each PID. However, in the mean time the
process might've finished.

Rather than create a fixed size Process slice and insert each value add a
fixed index, use a dynamically sized slice (with an initial capacity
that matches the likely final length).

If any of the PIDs fails to map to a process, we assume the process has
gone away and skip it.