Use gotip (pre 1.16) to compile a native executable for Apple Silicon
This adds a new build step that compiles a native agent for Apple
Silicon. I think. I don't have one to test on.

Cross compiling for Apple Silicon requires `GOARCH=arm64` and `GOOS=darwin`,
which is a valid combination on unreleased version of go and will be
valid in the upcoming 1.16 release.

We currently compile using go 1.15, so this new step uses gotip to
install the latest unreleased golang. That probably means the
darwin-arm64 binary it spits out should be considered experimental.

A native Apple Silicon executable isn't *that* important. The existing
darwin-amd64 executable we ship works just fine under rosetta2, but
native support is neat to have.