pin compilation of apple silicon binary to a golang CL that includes the fixes we need
We're nervous about always building the darwin-arm64 binary with golang
master. Compiling with a potentially different golang version each time
is a recipe for non-deterministic output and hard to debug regressions.

Ideally we'd like to pin to a specific golang master commit, but I can't
see a way to do that with gotip (see this open issue [1]). I think
pinning to this CL will get a version that has apple the fixes we're
after though.

[1] https://github.com/golang/go/issues/39650