compile the darwin/arm64 binary using 1.16beta1
In PR #1339 we started building a native binary for Apple M1 systems,
and we used gotip to compile with the latest master version of go.

A 1.16beta1 release is now available, and we can switch to compiling the
darwin/arm64 binary with that so we get consistent output.

Eventually, when 1.16 final is released we can use that for all
compilation and the special case for darwin/arm64 can go away.