Allow signal bootstrap sends to commands on cancel to be customized
Currently bootstrap sends a `SIGTERM` to commands when it, itself, is interrupted.

This adds `cancel-signal` which accepts signals in the form of SIGTERM, SIGHUP, etc.

This is the same config that `start` accepts, and is based on the PR that added that feature: https://github.com/buildkite/agent/pull/1041

Finally, `start` passes its `cancel-signal` to `bootstrap`, to prevent having to customize the `bootstrap-script`.