Explicitly set a working directory when we start the bootstrap
Currently we start the bootstrap process without an explicit working
directory. Presumably the operating system assigns it one, maybe the
same as the main agent process?

We've seen an example of a customer build where the default working
directory for the bootstrap didn't exist. This results in a fatal
exception at [1] and the job fail to run.

Our process abstraction supports setting an explicit working directory
for the child process [2], so we may as well use it.

I've opted to use the agent build path, which should be a directory that
exists. There's no real reason for the bootstrap process to start with that as
the working directory, other than the fact that it should exist.

Not long after it starts, the bootstrap process will find or create the full
checkout path where the source code will be cloned to, and then change its own
working directory to that [3].

Our process abstraction wraps cmd.Exec [4], which has this so say about
setting the working directory:

    // Dir specifies the working directory of the command.
    // If Dir is the empty string, Run runs the command in the
    // calling process's current directory.

[1] https://github.com/buildkite/agent/blob/bed05dda2d7bfb888d80c621af827ee2bce39144/bootstrap/shell/shell.go#L62-L65
[2] https://github.com/buildkite/agent/blob/bed05dda2d7bfb888d80c621af827ee2bce39144/process/process.go#L127-L132
[3] https://github.com/buildkite/agent/blob/bed05dda2d7bfb888d80c621af827ee2bce39144/bootstrap/bootstrap.go#L842-L846
[4] https://golang.org/pkg/os/exec/#Cmd