Use an updated version of our gopkg.in/yaml.v2 fork
We're were using a fork of gopkg.in/yaml.v2 (v2.1.1) with a patch to fix
ordering of Maps when used in conjunction with named anchors (this PR
[1]).

Upstream is now at v2.4.0, however the Maps+anchors patch hasn't been
merged. We've updated our fork to be v2.4.0+the patch, and this will
start using that new version.

The list of changes between v2.1.1 and v2.4.0 is fairly small [2], but
it includes a number of parsing bugs that could result in a panic, like
the 3 commits linked at [3]. I've confirmed this fixes a panic reported
by one of our customers while uploading a pipeline.yml.

[1] https://github.com/go-yaml/yaml/pull/331
[2] https://github.com/go-yaml/yaml/compare/v2.1.1...v2.4.0
[3] https://github.com/go-yaml/yaml/issues/323