Send queue metrics to Datadog when job received
We would like to also send how long a job was queued for when a job is
acquired by the agent to Datadog. This will empower teams to validate
that their clusters are not causing jobs to queue for "too long" (for
the users definition of that) on the assumption that all jobs will
eventually execute on an agent. Note that this :see_no_evil: that jobs
get stuck indefinitely.

This is not a complete Pull Request as we are waiting for our friends in
the Buildkite team to add the actual "scheduled_at" time to the acquire
job API, which will enable this pull request to actually calculate how
long the job was sitting in the queue for.