Option to send Datadog Metrics as Distribution
Datadog has added support for the
["Distribution"](https://docs.datadoghq.com/metrics/distributions/) type
of metric.

As a summary, the existing "timing" types in Datadog calculate things
like average, min, max, p95 on the client itself.  Therefore, it only
has visibility in what the client has access to. In the context of a
build job, this means that, in most cases, average == min == max == p95,
because they'll only be one job per flush interval.

A Distribution, instead, calculates all of these metrics server side at
the Datadog level itself. This makes it possible to aggregate metrics
across a wide variety of clients or sources. For example, you could have
a hundred Buildkite agents submit their job duration as raw data to
Datadog. Datadog will take the responsibility of calculating all the
rollup values based on the raw data.