Resolve the absolute path to the buildkite-agent binary for bootstrap
Previously, the built-in bootstrap subcommand was called using the raw value of `os.Args[0]`, which meant it could fail if the path was relative to the current working directory, as jobs run in a build-specific directory.

This patch updates it to use `filepath.Abs` to resolve an absolute version of that path, so, for example, calling `.buildkite-agent/bin/buildkite-agent` from the user's home directory would no longer fail to run jobs.

It also treats an error resolving this path as fatal, though looking at the code for it this should only really happen if the current working directory is deleted, but given this happens at agent start that is extremely unlikely.

Fixes #1439