Extract Uploader Client setup from ArtifactUploader
We'd like to be able to configure the S3 Uploader with the acl
permissions for each artifact, via the CLI. In order to plumb that down
to the S3 Client, I considered teaching ArtifactUploader about the ACL
types but it's a detail of the S3 Client and the ArtifactUploader
doesn't really need to care about it. There'll be other types of clients
we want to support and teaching the ArtifactUploader seems like it will
grow to know too many things.