Retry 500 responses when batch creating artifacts
Previously, when attempting to batch-create artifacts, if the agent API returned a 500 error, we would just sorta say 'okay' and never try again, emitting a somewhat confusing log message of:

```
POST https://agent.buildkite.com/v3/jobs/<some job id>/artifacts: 500 (Attempt 1/10 retrying in 10s)
``` ... and then a failure message with no retries.

This commit makes it so that when we run across a 500 when trying to batch-create some artifacts, we retry the operation, as the logs imply we should. also, 500s can be (though are not always) transient, so retrying them is a good idea regardless.

Closes #1565