update github.com/aws/aws-sdk-go to the latest v1 release
I'm not aware of any bugfixes or changes that we need urgently, but it's
approaching 2 years since we last upgraded (PR #1234) and we don't want
an important dep like this to trail too far behind. Mainly because the
longer the lag, the harder it might be to get back to current.

Once this merges I might investigate whether it's worth enabling
dependabot for go modules so the bumps can happen semi-regularly.