enable dependabot for go.mod
It would be nice to slowly bump the various dependencies we have, to
avoid situations like #1573 where a key dependency is nearly 2 years
out of date.

On the other hand, we don't want to be overwhelmed by tiny bumps all the
time.

Here's one we we could walk that path. Allow dependabot to open PRs with
updates, but not more than 2 per week. I'm not wedded to these values,
but I thought I'd start with a strawman.