Create directories only accessible by current user and group.
Note that the mode passed to mkdir is masked by umask.
The default umask of most systems is 0022 or at least 0002.

    $ echo "$(uname): $(umask)"
    Darwin: 022

    $ for img in archlinux debian ubuntu centos alpine; do echo -n "$img: "; docker run --rm $img sh -c umask; done
    archlinux: 0022
    debian: 0022
    ubuntu: 0022
    centos: 0022
    alpine: 0022

Given umask = 0022:

- The previous mode of 0777 created directories as: 0755 (drwxr-x---)
- The new mode of 0770 creates directories as:      0750 (drwxr-x---)

In neither case is the directory world-writable.

This patch drops the ability for other users to read the directory
regardless of the umask in effect during creation.

It continues to delegate group access control to umask, so that system
administrators can use that standard mechanism to control how different
users in the same group can share resources. The standard 0022 umask
provides a safe default of disallowing write access from other users in
the same group.