agent/plugin/definition_test.go(additionalProperties): Add tests.
I was messing around with this because for ages i could not work out why Buildkite wasn't throwing
an error when i passed in undefined additional config to a plugin i was using.  Eventually i worked
out that the `--[no-]plugin-validation` feature was introduced in
https://github.com/buildkite/agent/pull/748 back in 2018, but was never made default-on!  This was a
surprise to me, but i thought i might as well leave these tests here.

* Ensure additionalProperties = false complains when extra parameters are passed.

* Ensure additionalProperties = true allows extra parameters.